make conversions explicit and use std namespace for math functions

+++ core/src/graphics/colorImpl.h
@@ -2,9 +2,9 @@
 
 namespace spruce {
 	constexpr color::color(int32 hex) {
-		r = ((hex >> 24) & 0xFF) / 255.0f;
-		g = ((hex >> 16) & 0xFF) / 255.0f;
-		b = ((hex >> 8) & 0xFF) / 255.0f;
-		a = (hex & 0xFF) / 255.0f;
+		r = static_cast<float>((hex >> 24) & 0xFF) / 255.0f;
+		g = static_cast<float>((hex >> 16) & 0xFF) / 255.0f;
+		b = static_cast<float>((hex >> 8) & 0xFF) / 255.0f;
+		a = static_cast<float>(hex & 0xFF) / 255.0f;
 	}
 }
